Deprecate Flume.start/2 and Flume.start_link/0 functions
* Move it to Flume.Supervisor

* Move Flume APIs to Flume module and remove it from
  Flume.API
